+++ src/main/clojure/com/github/clojure_lsp/intellij/workspace_edit.clj
@@ -2,6 +2,7 @@
   (:require
    [com.github.clojure-lsp.intellij.client :as lsp-client]
    [com.github.clojure-lsp.intellij.editor :as editor]
+   [com.github.clojure-lsp.intellij.file-system :as file-system]
    [com.github.ericdallo.clj4intellij.app-manager :as app-manager]
    [com.github.ericdallo.clj4intellij.logger :as logger]
    [com.github.ericdallo.clj4intellij.util :as util]))
@@ -9,7 +10,7 @@
 (set! *warn-on-reflection* true)
 
 (defmethod lsp-client/workspace-apply-edit :default [{:keys [project]} {:keys [label edit]}]
-  (editor/apply-workspace-edit project label false edit)
+  (file-system/apply-workspace-edit project label false edit)
   {:applied true})
 
 (defmethod lsp-client/show-document :default [{:keys [project]} {:keys [uri take-focus selection]}]