+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased]
 
 - Support Create workspace edit operation. Fixes #51
+- Support Rename workspace edit operation
+- Support workspace/willRenameFiles allowing to rename ns, renaming the files properly.
 
 ## 2.5.0